extracted util.py
